requirements.txt
requests==2.26.0
django==2.2.16
djangorestframework==3.12.4
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
djangorestframework_simplejwt==4.8.0
django-filter==2.4.0
drf-yasg